reorganized folders and fix appropriate modal to pop up on add edit and delete
