+++ BeaconAndLoaves/ClientApp/src/components/ConfirmationModal/ConfrimationModal.js
