+++ BeaconAndLoaves/ClientApp/src/components/LikeButton/LikeButton.scss
