+++ BeaconAndLoaves/ClientApp/src/components/Rental/Rental.scss
