+++ BeaconAndLoaves/ClientApp/src/components/WarningModal/WarningModal.js
