+++ BeaconAndLoaves/ClientApp/src/components/WarningModal/WarningModal.scss
