--- BeaconAndLoaves/ClientApp/src/components/pages/EditProperty/EditProperty.scss
@@ -1,9 +1,0 @@
-.edit-property {
-    width: 60%;
-    margin-top: 5%;
-    color: white;
-}
-
-.edit-property-title {
-    color: white;
-}
