+++ BeaconAndLoaves/ClientApp/src/components/pages/Profile/Profile.js
@@ -4,7 +4,7 @@ import accountShape from '../../../helpers/propz/accountShape';
 import userRequests from '../../../helpers/data/userRequests';
 import WarningModal from '../WarningModal/WarningModal';
 import paymentMethodRequests from '../../../helpers/data/paymentMethodRequests';
-import SinglePaymentMethodModal from '../SinglePaymentMethodModal/SinglePaymentMethodModal';
+import SinglePaymentMethodModal from '../../SinglePaymentMethodModal/SinglePaymentMethodModal';
 import propertiesRequests from '../../../helpers/data/propertiesRequests';
 
 import './Profile.scss';
